requirements.txt
tensorflow>=2.15
keras>=3.3
scikit-learn>=1.3
matplotlib>=3.7
numpy>=1.24
pillow>=10.0

# backend
fastapi>=0.111
uvicorn[standard]>=0.29
python-multipart>=0.0.9